+++ DevilDaggersAssetCore/Headers/AbstractResourceHeader.cs
@@ -7,7 +7,7 @@ namespace DevilDaggersAssetCore.Headers
 		protected AbstractResourceHeader(byte[] headerBuffer)
 		{
 			Buffer = new byte[ByteCount];
-			Buf.BlockCopy(headerBuffer, 0, Buffer, 0, ByteCount);
+			Buf.BlockCopy(headerBuffer, 0, Buffer, 0, (int)ByteCount);
 		}
 	}
 }
